Rename RSC implementation of `NextIntlClientProvider` to have a unique name
