+++ packages/next-intl/src/react-server/index.tsx
@@ -13,7 +13,7 @@ export {default as useFormatter} from './useFormatter';
 export {default as useNow} from './useNow';
 export {default as useTimeZone} from './useTimeZone';
 export {default as useMessages} from './useMessages';
-export {default as NextIntlClientProvider} from './NextIntlClientProvider';
+export {default as NextIntlClientProvider} from './NextIntlClientProviderServer';
 
 // Everything from `core`
 export * from 'use-intl/core';
